+++ app/src/main/java/com/mskd/flux/screens/player/PlayerScreen.kt
@@ -117,6 +117,7 @@ fun PlayerScreen(
     }
 
     BackHandler(enabled = true) {
+        interfaceVisibilityCountdown = 3
         viewModel.handleIntent(PlayerIntent.OnBackTap)
     }