+++ shared/src/commonMain/kotlin/com/mskd/flux/core/database/data/Database.kt
@@ -31,7 +31,7 @@ import kotlinx.coroutines.Dispatchers
         GenreEntity::class,
         HistoryEntity::class
     ],
-    version = 9,
+    version = 11,
     autoMigrations = [
         AutoMigration(from = 1, to = 2),
         AutoMigration(from = 2, to = 3),
@@ -40,6 +40,8 @@ import kotlinx.coroutines.Dispatchers
         AutoMigration(from = 6, to = 7),
         AutoMigration(from = 7, to = 8),
         AutoMigration(from = 8, to = 9),
+        AutoMigration(from = 9, to = 10),
+        AutoMigration(from = 10, to = 11),
     ]
 )
 @TypeConverters(Converters::class)