+++ shared/src/commonMain/kotlin/com/mskd/flux/features/catalog/domain/model/CatalogPreferences.kt
@@ -10,4 +10,5 @@ data class CatalogPreferences(
     val sortingMode: CatalogSortingMode,
     val viewMode: CatalogViewMode,
     val token: String,
+    val privateFolderEnabled: Boolean,
 )