+++ shared/src/commonMain/kotlin/com/mskd/flux/features/files/domain/usecase/GetDeviceFilesUseCase.kt
@@ -18,5 +18,6 @@ class GetDeviceFilesUseCaseImpl(
             .map { source -> async { source.getFiles() } }
             .awaitAll()
             .flatten()
+            .distinctBy { it.path }
     }
 }