+++ CHANGELOG.md
@@ -2,5 +2,6 @@
 
 - UI: Season name and episodes count
 - UI: Animation for ArtworkDescriptionsPager with one page
+- UI: Item corner shape consistency
 - FIX: Last watched carousel image
 - FIX: Retrieve episodes above 99