+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 - UI: Items shape and count
 - CORE: Player compatibility 
 - CORE: All KMP 
+- FIX: Issue to show buttons in player with back button