+++ CHANGELOG.md
@@ -1,3 +1,5 @@
-## Version 1.8.0
+## Version 1.8.1
 
-- Feature: New History feature
+- Feature: Private folder
+- Feature: Fetch NSFW
+- Fix: Crash from 1.8.0